hooksの中身をリファクタリングした
