+++ nextjs/hooks/useAiGame.ts
@@ -1,30 +1,7 @@
-import { useState, useCallback, useEffect, useMemo, useRef } from "react";
+import { useState, useCallback, useEffect, useRef } from "react";
 import { useRouter } from "next/navigation";
-import {
-	PieceData,
-	INITIAL_BOARD,
-	PROMOTE_MAP,
-	DEMOTE_MAP,
-	HandPieces,
-} from "@/utils/shogiConstants";
-import {
-	getLegalMovesForPiece,
-	getLegalDrops,
-	boardFromPieces,
-	type LegalTarget,
-	type PieceInfo,
-} from "@/lib/shogi/board";
-import { useBoard } from "./useBoard";
-import { useHands } from "./useHands";
-
-// 打ち駒の漢字 → USI駒文字
-const DROP_KANJI_TO_USI: Record<string, string> = {
-	"歩": "P",
-	"銀": "S",
-	"金": "G",
-	"角": "B",
-	"飛": "R",
-};
+import { boardFromPieces, type PieceInfo } from "@/lib/shogi/board";
+import { useGameLogic } from "./useGameLogic";
 
 // USI駒文字 → 漢字
 const USI_TO_DROP_KANJI: Record<string, string> = {
@@ -40,23 +17,32 @@ export function useAiGame(
 	aiDepth: number = 4
 ) {
 	const router = useRouter();
-	const { board, movePiece, dropPiece } = useBoard(INITIAL_BOARD);
-	const { senteHand, goteHand, addCapturedPiece, removeHandPiece } = useHands();
 
-	const [selected, setSelected] = useState<{ row: number; col: number } | null>(null);
-	const [selectedHandPiece, setSelectedHandPiece] = useState<string | null>(null);
-	const [turn, setTurn] = useState<"sente" | "gote">("sente");
-	const [promoteDialog, setPromoteDialog] = useState<{
-		from: { row: number; col: number };
-		to: { row: number; col: number };
-	} | null>(null);
+	const {
+		board,
+		senteHand,
+		goteHand,
+		selected,
+		setSelected,
+		selectedHandPiece,
+		setSelectedHandPiece,
+		turn,
+		promoteDialog,
+		setPromoteDialog,
+		isMyTurn,
+		applyMove,
+		applyDrop,
+		isLegalTarget,
+		isLegalDropTarget,
+		getLegalTargetInfo,
+	} = useGameLogic(mySide);
+
 	const [aiThinking, setAiThinking] = useState(false);
 	const [gameOver, setGameOver] = useState<string | null>(null);
 
-	const isMyTurn = turn === mySide;
 	const aiSide = mySide === "sente" ? "gote" : "sente";
 
-	// 最新のboard/handをrefで保持（非同期コールバック用）
+	// 最新のboard/handをrefで保持（非同期AIコールバック用）
 	const boardRef = useRef(board);
 	const senteHandRef = useRef(senteHand);
 	const goteHandRef = useRef(goteHand);
@@ -65,73 +51,6 @@ export function useAiGame(
 	senteHandRef.current = senteHand;
 	goteHandRef.current = goteHand;
 	turnRef.current = turn;
-
-	// ========= 合法手計算 =========
-
-	const legalTargets: LegalTarget[] = useMemo(() => {
-		if (!selected) return [];
-		return getLegalMovesForPiece(
-			board as (PieceInfo | null)[][],
-			turn,
-			senteHand,
-			goteHand,
-			selected.row,
-			selected.col
-		);
-	}, [selected, board, turn, senteHand, goteHand]);
-
-	const legalDropTargets: { row: number; col: number }[] = useMemo(() => {
-		if (!selectedHandPiece) return [];
-		return getLegalDrops(
-			board as (PieceInfo | null)[][],
-			turn,
-			senteHand,
-			goteHand,
-			selectedHandPiece
-		);
-	}, [selectedHandPiece, board, turn, senteHand, goteHand]);
-
-	const isLegalTarget = useCallback(
-		(row: number, col: number) => legalTargets.some((t) => t.row === row && t.col === col),
-		[legalTargets]
-	);
-
-	const isLegalDropTarget = useCallback(
-		(row: number, col: number) => legalDropTargets.some((t) => t.row === row && t.col === col),
-		[legalDropTargets]
-	);
-
-	const getLegalTargetInfo = useCallback(
-		(row: number, col: number): LegalTarget | undefined =>
-			legalTargets.find((t) => t.row === row && t.col === col),
-		[legalTargets]
-	);
-
-	// ========= ローカル適用 =========
-
-	const applyMove = useCallback(
-		(from: { row: number; col: number }, to: { row: number; col: number }, promote: boolean) => {
-			const { capturedKanji, capturingSide } = movePiece(from, to, promote);
-			if (capturedKanji && capturingSide) {
-				addCapturedPiece(capturedKanji, capturingSide);
-			}
-			setTurn((prev) => (prev === "sente" ? "gote" : "sente"));
-			setSelected(null);
-			setPromoteDialog(null);
-		},
-		[movePiece, addCapturedPiece]
-	);
-
-	const applyDrop = useCallback(
-		(kanji: string, to: { row: number; col: number }, side: "sente" | "gote") => {
-			dropPiece(kanji, to, side);
-			removeHandPiece(kanji, side);
-			setTurn((prev) => (prev === "sente" ? "gote" : "sente"));
-			setSelectedHandPiece(null);
-			setSelected(null);
-		},
-		[dropPiece, removeHandPiece]
-	);
 
 	// ========= プレイヤーアクション =========
 
@@ -190,7 +109,6 @@ export function useAiGame(
 	// AIの手番になったら自動で思考開始
 	useEffect(() => {
 		if (turn === aiSide && !gameOver && !aiThinking) {
-			// 少し遅延を入れてUIが更新されてから思考開始
 			const timer = setTimeout(() => {
 				requestAiMove();
 			}, 300);
@@ -251,6 +169,7 @@ export function useAiGame(
 			board, selected, selectedHandPiece, mySide, turn, isMyTurn,
 			aiThinking, gameOver, promoteDialog,
 			isLegalDropTarget, getLegalTargetInfo, executeMove, applyDrop,
+			setSelected, setSelectedHandPiece, setPromoteDialog
 		]
 	);
 
@@ -263,7 +182,7 @@ export function useAiGame(
 			setSelected(null);
 			setSelectedHandPiece((prev) => (prev === kanji ? null : kanji));
 		},
-		[isMyTurn, aiThinking, gameOver, promoteDialog, turn, mySide]
+		[isMyTurn, aiThinking, gameOver, promoteDialog, turn, mySide, setSelected, setSelectedHandPiece]
 	);
 
 	const handleEndMatch = useCallback(() => {
