+++ nextjs/hooks/useShogiGame.ts
@@ -1,121 +1,37 @@
-import { useState, useCallback, useEffect, useMemo } from "react";
+import { useCallback, useEffect } from "react";
 import { useRouter } from "next/navigation";
 import { Socket } from "socket.io-client";
-import {
-	PieceData,
-	INITIAL_BOARD,
-	HandPieces,
-} from "@/utils/shogiConstants";
-import {
-	getLegalMovesForPiece,
-	getLegalDrops,
-	type LegalTarget,
-	type PieceInfo,
-} from "@/lib/shogi/board";
-import { useBoard } from "./useBoard";
-import { useHands } from "./useHands";
+import { useGameLogic } from "./useGameLogic";
 
 export function useShogiGame(
 	socket: Socket | null | undefined,
 	roomId: string | undefined,
 	mySide: "sente" | "gote",
 	wsStatus: "connected" | "disconnected" | "connecting"
-) 
-{
+) {
 	const router = useRouter();
 
-	// 子フック
-	const { board, movePiece, dropPiece } = useBoard(INITIAL_BOARD);
-	const { senteHand, goteHand, addCapturedPiece, removeHandPiece } = useHands();
+	const {
+		board,
+		turn,
+		senteHand,
+		goteHand,
+		selected,
+		setSelected,
+		selectedHandPiece,
+		setSelectedHandPiece,
+		promoteDialog,
+		setPromoteDialog,
+		isMyTurn,
+		isLegalTarget,
+		isLegalDropTarget,
+		getLegalTargetInfo,
+		applyMove,
+		applyDrop,
+	} = useGameLogic(mySide);
 
-	// UI状態
-	const [selected, setSelected] = useState<{ row: number; col: number } | null>(null);
-	const [selectedHandPiece, setSelectedHandPiece] = useState<string | null>(null);
-	const [turn, setTurn] = useState<"sente" | "gote">("sente");
-	const [promoteDialog, setPromoteDialog] = useState<{
-		from: { row: number; col: number };
-		to: { row: number; col: number };
-	} | null>(null);
+	// ========= アクション (WebSocket付き) =========
 
-	const isMyTurn = turn === mySide;
-
-	// ========= 合法手計算 =========
-
-	const legalTargets: LegalTarget[] = useMemo(() => {
-		if (!selected) return [];
-		return getLegalMovesForPiece(
-			board as (PieceInfo | null)[][],
-			turn,
-			senteHand,
-			goteHand,
-			selected.row,
-			selected.col
-		);
-	}, [selected, board, turn, senteHand, goteHand]);
-
-	const legalDropTargets: { row: number; col: number }[] = useMemo(() => {
-		if (!selectedHandPiece) return [];
-		return getLegalDrops(
-			board as (PieceInfo | null)[][],
-			turn,
-			senteHand,
-			goteHand,
-			selectedHandPiece
-		);
-	}, [selectedHandPiece, board, turn, senteHand, goteHand]);
-
-	const isLegalTarget = useCallback(
-		(row: number, col: number) => {
-			return legalTargets.some((t) => t.row === row && t.col === col);
-		},
-		[legalTargets]
-	);
-
-	const isLegalDropTarget = useCallback(
-		(row: number, col: number) => {
-			return legalDropTargets.some((t) => t.row === row && t.col === col);
-		},
-		[legalDropTargets]
-	);
-
-	const getLegalTargetInfo = useCallback(
-		(row: number, col: number): LegalTarget | undefined => {
-			return legalTargets.find((t) => t.row === row && t.col === col);
-		},
-		[legalTargets]
-	);
-
-	// ========= アクション =========
-
-	// ローカルのみ適用（WebSocket送信なし）
-	const applyMove = useCallback(
-		(from: { row: number; col: number }, to: { row: number; col: number }, promote: boolean) => {
-			const { capturedKanji, capturingSide } = movePiece(from, to, promote);
-
-			if (capturedKanji && capturingSide) {
-				addCapturedPiece(capturedKanji, capturingSide);
-			}
-
-			setTurn((prev) => (prev === "sente" ? "gote" : "sente"));
-			setSelected(null);
-			setPromoteDialog(null);
-		},
-		[movePiece, addCapturedPiece]
-	);
-
-	const applyDrop = useCallback(
-		(kanji: string, to: { row: number; col: number }, side: "sente" | "gote") => {
-			dropPiece(kanji, to, side);
-			removeHandPiece(kanji, side);
-
-			setTurn((prev) => (prev === "sente" ? "gote" : "sente"));
-			setSelectedHandPiece(null);
-			setSelected(null);
-		},
-		[dropPiece, removeHandPiece]
-	);
-
-	// ローカル適用 ＋ WebSocket送信
 	const executeMove = useCallback(
 		(from: { row: number; col: number }, to: { row: number; col: number }, promote: boolean) => {
 			applyMove(from, to, promote);
@@ -139,7 +55,7 @@ export function useShogiGame(
 		[turn, applyDrop, socket, roomId]
 	);
 
-	// ========= 相手の手を受信（WebSocket送信しない） =========
+	// ========= 相手の手を受信 =========
 
 	useEffect(() => {
 		if (!socket) return;
@@ -151,7 +67,6 @@ export function useShogiGame(
 			drop?: string;
 		}) => {
 			if (data.drop) {
-				// 相手の手番を推定（自分の逆）
 				const oppSide = mySide === "sente" ? "gote" : "sente";
 				applyDrop(data.drop, data.to, oppSide);
 			} else if (data.from) {
@@ -175,7 +90,7 @@ export function useShogiGame(
 
 			const cell = board[row][col];
 
-			// 持ち駒選択中 → 打ち / 解除
+			// 持ち駒選択中
 			if (selectedHandPiece) {
 				if (isLegalDropTarget(row, col)) {
 					executeDrop(selectedHandPiece, { row, col });
@@ -190,13 +105,11 @@ export function useShogiGame(
 
 			// 盤上の駒選択中
 			if (selected) {
-				// 自駒クリック → 選び直し
 				if (cell && cell.side === mySide) {
 					setSelected({ row, col });
 					return;
 				}
 
-				// 合法手チェック
 				const targetInfo = getLegalTargetInfo(row, col);
 				if (targetInfo) {
 					const from = { row: selected.row, col: selected.col };
@@ -224,6 +137,7 @@ export function useShogiGame(
 			board, selected, selectedHandPiece, mySide, turn, isMyTurn,
 			roomId, wsStatus, promoteDialog,
 			isLegalDropTarget, getLegalTargetInfo, executeMove, executeDrop,
+			setSelected, setSelectedHandPiece, setPromoteDialog
 		]
 	);
 
@@ -237,14 +151,12 @@ export function useShogiGame(
 			setSelected(null);
 			setSelectedHandPiece((prev) => (prev === kanji ? null : kanji));
 		},
-		[roomId, wsStatus, isMyTurn, promoteDialog, turn, mySide]
+		[roomId, wsStatus, isMyTurn, promoteDialog, turn, mySide, setSelected, setSelectedHandPiece]
 	);
 
 	const handleEndMatch = useCallback(() => {
 		router.push("/result" + (roomId ? `?roomId=${roomId}` : ""));
 	}, [router, roomId]);
-
-	// ========= 返り値 =========
 
 	return {
 		board,
